+++ hashmap/hashmap.h
@@ -11,16 +11,16 @@ typedef struct hashmap hashmap;
 
 hashmap *hashmap_create(
     int (*)(const void *, const void *),
-    size_t (*)(const void *),
+    unsigned long (*)(const void *),
     size_t
 );
 void *hashmap_insert(hashmap *, void *);
 void *hashmap_delete(hashmap *, const void *);
 void *hashmap_search(hashmap *, const void *);
 void hashmap_destroy(hashmap *, void (*)(void *));
-size_t hashmap_hash_str(unsigned char *);
-size_t hashmap_hash_uint(unsigned int);
-size_t hashmap_hash_ulong(unsigned long);
+unsigned long hashmap_hash_str(unsigned char *);
+unsigned long hashmap_hash_uint(unsigned int);
+unsigned long hashmap_hash_ulong(unsigned long);
 
 #endif /* HASHMAP_H */
